+++ Adj/Program.hs
@@ -1,7 +1,5 @@
 module Adj.Program (module Exports) where
 
 import Adj.Program.Generation as Exports
-import Adj.Program.Instruction as Exports
-import Adj.Program.Construction as Exports
 import Adj.Program.Option as Exports
 import Adj.Program.Progress as Exports